+++ data-warehouse/__init__.py
@@ -1,0 +1,1 @@
+# data-warehouse package v0.1.0 (2025-08-19)